cvar: add `cl_maxfps` which replaces `com_maxfps`
Add a new client `cl_maxfps` which allows specifying the maximum desired fps
value; which may not be the same as the maximum allowed.

`com_maxfps` is now derived from `cl_maxfps`.

Current values are 250 when `pmove_fixed=1` and 125 when `pmove_fixed=0`

The default value of `cl_maxfps -1` should probably be preferred because it
just picks the best currently supported max.  Future patches will raise this to
500 in the pmove_fixed=1 case.